create interface for search algorithm to improve readability
