+++ src/main/kotlin/logic/search/SearchByName.kt
@@ -1,0 +1,6 @@
+package org.berlin.logic.search
+
+interface SearchByName {
+    fun search(textToSearchIn: String,wordToSearchBy: String): Boolean
+
+}
